+++ app/users.js
@@ -134,14 +134,14 @@ router.delete('/:id', async (req, res) => {
     return;
   }
 
-  //Cascade delete all the operations done on the car
-  Operation.deleteMany({ employee: user._id }, (err) => {
-    if (err) {
-      console.log(err);
-      res.status(404).send('Cannot cascade delete the operation');
-      return;
-    }
-  });
+  //Check if the user is associated with any operations, in that case it cannot be deleted
+  let operations = await Operation.find({ employee: user._id });
+
+  if(operations.length != 0) {
+    res.status(403).send();
+    console.log('Cannot delete the user, it has some operations associated to it');
+    return;
+  }
 
   await user.deleteOne();
   console.log('user removed');
